APPEALS-50018: updated feature toggles, all functions, completed json, added route
